backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pillow==11.1.0
numpy==2.2.2
python-multipart==0.0.20
aiofiles==24.1.0
